% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD.R
\name{CFD}
\alias{CFD}
\title{CFD}
\usage{
CFD(d, s = 1)
}
\arguments{
\item{d}{file path not raw vector with pdf data}

\item{s}{row of specCFD 1:default(current version)}
}
\description{
証拠金入出金明細書(.pdf)を(.csv)に変換する。
}
