% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sbi.R
\name{read_sbi}
\alias{read_sbi}
\title{read_sbi}
\usage{
read_sbi(d = getwd(), f = ".pdf$", p = "", t = "")
}
\arguments{
\item{d}{file path}

\item{f}{file name}

\item{p}{pattern file ex."^inf"}

\item{t}{title}
}
\description{
read_sbi
}
