% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specCFD.R
\docType{data}
\name{specCFD.init}
\alias{specCFD.init}
\title{Datasets of specCFD}
\format{
A data frame with 1or2 rows and 4 variables:
}
\usage{
specCFD.init()
}
\description{
A dataset of spec of pdf files in CFD.
}
\keyword{datasets}
