% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD-pdf.R, R/read_sbi-sub.R, R/sbitools.R
\docType{package}
\name{sbitools}
\alias{sbitools-package}
\alias{sbitools}
\alias{CFD.pdf}
\alias{ynx}
\title{CFD.pdf}
\usage{
CFD.pdf(specCFD, df, p, s = 1)

ynx(df_p, n, s = 0, e = Inf)
}
\arguments{
\item{specCFD}{settings}

\item{df}{pdf_data}

\item{p}{pages}

\item{s}{pdf_data[[n]]$"x" Starting position}

\item{df_p}{pdf_data[[n]]}

\item{n}{pdf_data[[n]]$"y"}

\item{e}{pdf_data[[n]]$"x" End position}
}
\description{
Put csv from pdf
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qiita.com/nt1969m/items/69ea8d0957c1fca7763f (CFD) https://qiita.com/nt1969m/items/890846d529d08c8edbb2 (As Allocated sharesの略のつもり) https://qiita.com/nt1969m/items/8e9d6f3454c2087bf6ec (Div配当金)}
}

}
\author{
\strong{Maintainer}: Naohito Takahashi \email{nt1969m@gmail.com}

}
