% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD-pdf.R, R/CFD.R, R/sbitools.R
\docType{package}
\name{sbitools}
\alias{sbitools}
\alias{CFD.pdf}
\alias{CFD}
\title{CFD.pdf}
\usage{
CFD.pdf(specCFD, df, p, s = 1)

CFD(d, s = 1)
}
\arguments{
\item{specCFD}{settings}

\item{df}{pdf_data}

\item{p}{pages}

\item{s}{row of specCFD 1:default(current version)}

\item{d}{file path not raw vector with pdf data}
}
\description{
説明

複数の(.pdf)を(.csv)に変換する。

Put csv from pdf
}
