% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFD-pdf.R, R/sbitools.R
\docType{package}
\name{sbitools}
\alias{sbitools}
\alias{CFD.pdf}
\title{CFD.pdf}
\usage{
CFD.pdf(specCFD, df, p, s = 1)
}
\arguments{
\item{specCFD}{settings}

\item{df}{pdf_data}

\item{p}{pages}

\item{s}{row of specCFD}
}
\description{
説明

Put csv from pdf
}
